edit somethings

+++ BookingTicketOnline/Models/Revenue.cs
@@ -7,7 +7,7 @@ namespace BookingTicketOnline.Models
     {
         public int Id { get; set; }
         public int? PaymentId { get; set; }
-        public decimal? TotalRevenue { get; set; }
+        public int? TotalRevenue { get; set; }
         public DateTime? FromDate { get; set; }
         public DateTime? ToDate { get; set; }
